backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
torch==2.8.0
torchvision==0.23.0
opencv-python-headless==4.8.1.78
mediapipe==0.10.14
numpy==1.26.4
pillow==11.3.0
facenet-pytorch==2.5.3
tqdm==4.67.1
scikit-learn==1.6.1
